Fix test: verify V2 response structure and principal
